Fix a bug in lineage that stored erroneous data for new terms (#110861)
The problem was in _lineage_get_parent_lineage, which seems to never have
worked. All terms were given a depth of 0, and only fixed when
lineage_update_all was called. I've also added a call to lineage_update_all in
the install file so that previously broken installs will be fixed.